+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [next] 
 - Adding Component#onDestroy
 - Adding Keyboard events API
+- Adding Box2DGame, an extension of BaseGame to simplify lifecycle of Box2D components
 - Add onAnimateComplete for Animation (thanks @diegomgarcia)
 
 ## 0.18.2